feat: criar politica de autorizacao para usuario Admin

+++ app/Http/Controllers/BrandController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\BrandUpdateRequest;
 use App\Models\Brand;
 use App\Services\BrandServices;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BrandController extends Controller
 {
@@ -18,6 +19,8 @@ class BrandController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Brand::class);
+
         $brands = $this->brandServices->list();
 
         return response()->json($brands);
@@ -28,6 +31,8 @@ class BrandController extends Controller
      */
     public function store(BrandStoreRequest $request)
     {
+        Gate::authorize('create', Brand::class);
+
         $brand = $this->brandServices->store($request);
 
         return response()->json($brand);
@@ -38,6 +43,8 @@ class BrandController extends Controller
      */
     public function show(Brand $brand)
     {
+        Gate::authorize('view', $brand);
+
         return response()->json($brand);
     }
 
@@ -46,6 +53,8 @@ class BrandController extends Controller
      */
     public function update(BrandUpdateRequest $request, Brand $brand)
     {
+        Gate::authorize('update', $brand);
+
         $brand = $this->brandServices->update($request, $brand);
         
         return response()->json($brand);
@@ -56,6 +65,8 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand)
     {
+        Gate::authorize('delete', $brand);
+
         $this->brandServices->destroy($brand);
 
         return response()->json(["message" => "Successfully deleted"]);
